+++ corlett.h
@@ -55,7 +55,7 @@ typedef int corlett_lib_callback_t(int libnum, uint8 *lib, uint64 size, corlett_
 int corlett_decode(uint8 *input, uint32 input_len, corlett_t *c, corlett_lib_callback_t *lib_callback);
 void corlett_free(corlett_t *c);
 int corlett_tag_recognize(corlett_t *c, const char **target_value, int tag_num, const char *key);
-void corlett_length_set(uint32 length_ms, int32 fade_ms);
+void corlett_length_set(double length_seconds, double fade_seconds);
 uint32 corlett_sample_count(void);
 void corlett_sample_fade(stereo_sample_t *sample);
-uint32 psfTimeToMS(const char *str);
+double psfTimeToSeconds(const char *str);